AspNetCore.Testing readme, namespaces
