+++ src/Shared/DiagnosticIds/Experiments.cs
@@ -31,4 +31,5 @@ internal static class Experiments
     internal const string DocumentDb = "EXTEXP0011";
     internal const string AutoActivation = "EXTEXP0012";
     internal const string HttpLogging = "EXTEXP0013";
+    internal const string AspNetTesting = "EXTEXP0014";
 }
